pyproject.toml
[tool.poetry]
name = "telegram-forward-blocker"
version = "0.1.0"
description = ""
authors = ["coryphoenixxx <coryphoenixxx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
telethon = "^1.29.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
